requirements.txt
streamlit
supabase
supabase-py
pandas
plotly
pycountry
Pillow